refactor: GlobalAdvice 추가
