+++ src/main/java/roomescape/controller/advice/GlobalAdvice.java
@@ -1,0 +1,22 @@
+package roomescape.controller.advice;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import roomescape.exception.BadRequestException;
+import roomescape.exception.NotFoundException;
+
+@ControllerAdvice
+public class GlobalAdvice {
+
+    @ExceptionHandler(NotFoundException.class)
+    public ResponseEntity<String> notFoundExceptionHandler(NotFoundException exception) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(exception.getMessage());
+    }
+
+    @ExceptionHandler(BadRequestException.class)
+    public ResponseEntity<String> badRequestExceptionHandler(BadRequestException exception) {
+        return ResponseEntity.badRequest().body(exception.getMessage());
+    }
+}
